Complilation fix

+++ sys/list.h
@@ -19,10 +19,10 @@
 #include ARCH
 
 
-extern void lib_listAdd(void **list, void *t, u32 next_off, u32 prev_off);
+extern void lib_listAdd(void **list, void *t, size_t next_off, size_t prev_off);
 
 
-extern void lib_listRemove(void **list, void *t, u32 next_off, u32 prev_off);
+extern void lib_listRemove(void **list, void *t, size_t next_off, size_t prev_off);
 
 
 #define LIST_ADD_EX(list, t, next, prev) \
